moved e2e tests
