+++ pubky-testnet/Cargo.toml
@@ -22,6 +22,10 @@ url = "2.5.4"
 pubky = { path = "../pubky" }
 pubky-common = { path = "../pubky-common" }
 pubky-homeserver = { path = "../pubky-homeserver" }
+bytes = "1.10.1"
+
 
 [dev-dependencies]
 tracing-subscriber = "0.3.19"
+pkarr = "3.6.0"
+reqwest = "0.12.15"
