+++ pubky-testnet/src/e2e_tests/mod.rs
@@ -1,0 +1,3 @@
+mod auth;
+mod http;
+mod public;
